Code for Sieve of Eratosthenes
Contains the main method and two methods:
sieveAlgorithm finds the number of primes between a range of two numbers. It uses a bool of arrays and indexes through each element, marking an element true if the index is a prime number; false, if not.
validateStart verifies the starting range. The starting number must be equal to or greater than 1.